Added Servo Menu + Tests
Added a Servo Menu with manual and automatic tests.
Added ultrasonic sensor test, which will constantly update the position read in until the user clicks the encoder.
Added an "OK" confirmation to test screens to allow the user to move forward to the test or back to the main menu.